+++ packages/stream_chat_persistence/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 - Reintroduce support for experimental indexedDB on Web.
 
+🐞 Fixed
+
+- Fix offline message pagination.
+
 ## 5.0.0
 
 - Included the changes from version [4.3.0](#430) and [4.4.0](#440).